# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PetRetriever API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type PetGender = (typeof PetGender)[keyof typeof PetGender];

export const PetGender = {
  Male: "Male",
  Female: "Female",
  Unknown: "Unknown",
} as const;

export type PetStatus = (typeof PetStatus)[keyof typeof PetStatus];

export const PetStatus = {
  Verified: "Verified",
  Pending: "Pending",
  Incomplete: "Incomplete",
} as const;

export interface Pet {
  id: string;
  petId: string;
  name: string;
  species: string;
  breed: string;
  age: number;
  gender: PetGender;
  rhinariumId?: string | null;
  photoUrl?: string | null;
  status: PetStatus;
  createdAt: string;
}

export interface Owner {
  id: string;
  name: string;
  phone: string;
  email?: string | null;
}

export type PetWithOwner = Pet & {
  owner?: Owner;
};

export interface VaccinationRecord {
  id: string;
  petId: string;
  type: string;
  date: string;
  verified: boolean;
  notes?: string | null;
}

export interface MedicalRecord {
  id: string;
  petId: string;
  notes: string;
  createdAt: string;
}

export type PetProfile = PetWithOwner & {
  vaccinations?: VaccinationRecord[];
  medicalRecords?: MedicalRecord[];
};

export interface AdminStats {
  total: number;
  verified: number;
  pending: number;
  incomplete: number;
  recentRegistrations: number;
}

export type CreatePetRequestGender =
  (typeof CreatePetRequestGender)[keyof typeof CreatePetRequestGender];

export const CreatePetRequestGender = {
  Male: "Male",
  Female: "Female",
  Unknown: "Unknown",
} as const;

export interface CreatePetRequest {
  name: string;
  species: string;
  breed: string;
  age: number;
  gender: CreatePetRequestGender;
  rhinariumId?: string | null;
  photoUrl?: string | null;
  ownerName: string;
  ownerPhone: string;
  ownerEmail?: string | null;
}

export type UpdatePetRequestGender =
  (typeof UpdatePetRequestGender)[keyof typeof UpdatePetRequestGender];

export const UpdatePetRequestGender = {
  Male: "Male",
  Female: "Female",
  Unknown: "Unknown",
} as const;

export type UpdatePetRequestStatus =
  (typeof UpdatePetRequestStatus)[keyof typeof UpdatePetRequestStatus];

export const UpdatePetRequestStatus = {
  Verified: "Verified",
  Pending: "Pending",
  Incomplete: "Incomplete",
} as const;

export interface UpdatePetRequest {
  name?: string;
  species?: string;
  breed?: string;
  age?: number;
  gender?: UpdatePetRequestGender;
  rhinariumId?: string | null;
  photoUrl?: string | null;
  status?: UpdatePetRequestStatus;
}

export interface CreateVaccinationRequest {
  type: string;
  date: string;
  verified: boolean;
  notes?: string | null;
}

export interface CreateMedicalRequest {
  notes: string;
}

export interface VetLoginRequest {
  username: string;
  password: string;
}

export interface VetProfile {
  id: string;
  username: string;
  name: string;
  clinic?: string | null;
}

export interface VetLoginResponse {
  token: string;
  vet: VetProfile;
}

export interface UploadPhotoRequest {
  filename: string;
  dataUrl: string;
}

export interface UploadPhotoResponse {
  url: string;
}

export type ListPetsParams = {
  status?: ListPetsStatus;
  search?: string;
};

export type ListPetsStatus =
  (typeof ListPetsStatus)[keyof typeof ListPetsStatus];

export const ListPetsStatus = {
  Verified: "Verified",
  Pending: "Pending",
  Incomplete: "Incomplete",
} as const;

export type VerifyPetParams = {
  petId?: string;
  phone?: string;
  rhinariumId?: string;
};