# lib/api-zod/src/generated/types/vetLoginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PetRetriever API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VetLoginRequest {
  username: string;
  password: string;
}