# lib/api-zod/src/generated/types/vetLoginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PetRetriever API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VetProfile } from "./vetProfile";

export interface VetLoginResponse {
  token: string;
  vet: VetProfile;
}