# lib/api-zod/src/generated/types/vetProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PetRetriever API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VetProfile {
  id: string;
  username: string;
  name: string;
  clinic?: string | null;
}